+++ logix-emulator-module/web-ui/src/pages/PLCUpload/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./PLCUpload";
-export { default as PLCUpload } from "./PLCUpload";