+++ logix-emulator-module/web-ui/src/pages/TagBrowser/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./TagBrowser";
-export { default as TagBrowser } from "./TagBrowser";